feat(validation): now is not allowing to delete super administrators.

+++ src/context/api/userProfile/infraestructure/http-api/delete-userprofile-by-id/delete-by-id.controller.ts
@@ -30,6 +30,9 @@ export class DeleteUserProfileByIdController {
     @Delete(PrivateEndpoints.DELETE_PROFILE)
     async run(@Param('id', ParseIntPipe) id: number, @GetInformationByToken() gibt: any): Promise<void> {
         try {
+
+            if(id === 1) throw new BadRequestException('You cannot delete this profile');
+            
             const params = new DeleteByIdHttpDto();
             params.id = id;
             await this.deleteUserProfileByIdUseCase.execute(params);
